+++ journal/frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from '@emotion/react'
 import { CssBaseline, createTheme } from '@mui/material'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import './App.css'
-import { Contacts } from './features/contacts/Contacts'
+import { Contacts } from './features/people/PeopleCards'
 import { Nav } from './features/nav/Nav'
 
 const darkTheme = createTheme({
