--- journal/frontend/src/features/contacts/Contact.ts
@@ -1,8 +1,0 @@
-import { useLocalStorage } from "../storage/local";
-
-export type Contact = {
-  id: number;
-  name: string;
-  notes?: string;
-  avatar?: string;
-};
